+++ src/classes/api/index.test.ts
@@ -86,8 +86,11 @@ describe('API Class Tests', () => {
         expect(await API.addNote('', '', {year: 0, month: 0, day: 0, hour: 0, minute: 0, seconds: 0}, {year: 0, month: 0, day: 0, hour: 0, minute: 0, seconds: 0}, true, 0, [])).toBe(null);
         expect(NManager.createNote).not.toHaveBeenCalled();
 
-        expect(await API.addNote('', '', {year: 0, month: 0, day: 0, hour: 0, minute: 0, seconds: 0}, {year: 0, month: 0, day: 0, hour: 0, minute: 0, seconds: 0}, true, 0, [], '')).toBe(null);
+        expect(await API.addNote('', '', {year: 0, month: 0, day: 0, hour: 0, minute: 0, seconds: 0}, {year: 0, month: 0, day: 0, hour: 0, minute: 0, seconds: 0}, true, 0, [], '', 'asd')).toBe(null);
         expect(NManager.createNote).toHaveBeenCalledTimes(1);
+
+        expect(await API.addNote('', '', {year: 0, month: 0, day: 0, hour: 0, minute: 0, seconds: 0}, {year: 0, month: 0, day: 0, hour: 0, minute: 0, seconds: 0}, true, 0, [], '')).toBe(null);
+        expect(NManager.createNote).toHaveBeenCalledTimes(2);
     });
 
     test('Advance Time to Preset', () => {
@@ -293,12 +296,12 @@ describe('API Class Tests', () => {
         jest.spyOn(console, 'warn').mockImplementation(() => {});
         updateMigrationApplication(new MigrationApp());
         jest.spyOn(MigrationApplication, 'run').mockImplementation(async () => {});
-        expect(API.runMigration()).rejects.toBeUndefined();
+        expect(API.runMigration()).toBeUndefined();
 
         //@ts-ignore
         game.user.isGM = true;
         SC.primary = true;
-        expect(await API.runMigration()).toBeUndefined();
+        expect(API.runMigration()).toBeUndefined();
     });
 
     test('Seconds To Interval', () => {
@@ -321,6 +324,8 @@ describe('API Class Tests', () => {
 
     test('Show Calendar', () => {
         jest.spyOn(MainApplication, 'showApp').mockImplementation(() => {});
+        jest.spyOn(MainApplication, 'updateApp').mockImplementation(() => {});
+        jest.spyOn(MainApplication, 'rendered', 'get').mockReturnValueOnce(false).mockReturnValue(true)
         API.showCalendar();
         expect(console.error).toHaveBeenCalledTimes(1);
         expect(MainApplication.showApp).toHaveBeenCalledTimes(1);
@@ -329,15 +334,18 @@ describe('API Class Tests', () => {
         //@ts-ignore
         API.showCalendar({year: 'asd'}, false, '');
         expect(console.error).toHaveBeenCalledTimes(2);
-        expect(MainApplication.showApp).toHaveBeenCalledTimes(2);
+        expect(MainApplication.showApp).toHaveBeenCalledTimes(1);
+        expect(MainApplication.updateApp).toHaveBeenCalledTimes(2);
 
         tCal.months[0].current = true;
         tCal.months[0].days[0].current = true;
         API.showCalendar({month: 1});
-        expect(MainApplication.showApp).toHaveBeenCalledTimes(3);
+        expect(MainApplication.showApp).toHaveBeenCalledTimes(1);
+        expect(MainApplication.updateApp).toHaveBeenCalledTimes(4);
 
         API.showCalendar({year: 2020, month: -1, day: -1});
-        expect(MainApplication.showApp).toHaveBeenCalledTimes(4);
+        expect(MainApplication.showApp).toHaveBeenCalledTimes(1);
+        expect(MainApplication.updateApp).toHaveBeenCalledTimes(6);
     });
 
     test('Start Clock', () => {